+++ RELEASENOTES.md
@@ -1,1 +1,3 @@
+- Add a warning to help users who accidentally end up in the "running on the Bitbucket instance" flow in `bbs2gh migrate-repo`
 - Fix version check so that a version later than the latest version published on GitHub isn't considered old
+- Fix `download-logs` commands suggested by `bbs2gh migrate-repo` and `ado2gh migrate-repo` to use correct arguments